KafkaConsumer: move ConsumptionStrategy to `init`
Motivation:

One of our internal adopters asked for this as they were using mutliple
`KafkaConsumer`s that all shared the same `KafkaConsumerConfiguration`
but had different `ConsumptionStrategy`s. This meant they had to
duplicate their `KafkaConsumerConfiguration` just to accomodate for
different consumption strategies.

Modifications:

* move `ConsumptionStrategy` for `KafkaConsumer` from
  `KafkaConsumerConfiguration` to `KafkaConsumer.init`
* update `README`
* update tests
